feat: move old code to LegacyNightModeHelper and add logic to pick treatment
